collect call logic bug

+++ src/ripple/app/hook/applyHook.h
@@ -242,8 +242,8 @@ namespace hook
 
     uint32_t maxHookChainLength(void);
 
-    uint32_t computeExecutionFee(uint64_t instructionCount);
-    uint32_t computeCreationFee(uint64_t byteCount);
+    int64_t computeExecutionFee(uint64_t instructionCount);
+    int64_t computeCreationFee(uint64_t byteCount);
 
     struct HookResult
     {
